More palatable async command support
This approach removes the downsides of the parallel
`AsyncParsableCommand` protocol, with the downside of a bit of
boilerplate in the root type of the command tree.

In this formulation, `ParsableCommand` still provides only a non-async
version of `static func main`, with an async version provided by an
`AsyncMain` protocol. Instead of marking the root type of a command
hierarchy with @main, a developer will need to instead create a type
that conforms to AsyncMain and specifies the root command, like so:

@main enum Main: AsyncMain {
  typealias Command = <#command#>
}

The command hierarchy can then include async and non-async commands at
any level, including the root.